+++ src/storage/trend_state_models.py
@@ -1,7 +1,6 @@
 """Versioned live-state models for trend activation lineage."""
 
-from __future__ import annotations
-
+from __future__ import annotations  # noqa: I001
 from datetime import datetime
 from typing import TYPE_CHECKING, Any
 from uuid import UUID, uuid4
@@ -17,11 +16,11 @@ from sqlalchemy import (
     func,
     text,
 )
-from sqlalchemy.dialects.postgresql import JSONB
-from sqlalchemy.dialects.postgresql import UUID as PGUUID
+from sqlalchemy.dialects.postgresql import JSONB, UUID as PGUUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from src.storage.base import Base
+from src.storage.decimal_utils import Decimal  # noqa: TC001
 from src.storage.scoring_contract import (
     TREND_SCORING_MATH_VERSION,
     TREND_SCORING_PARAMETER_SET,
@@ -107,9 +106,9 @@ class TrendStateVersion(Base):
         server_default=text(f"'{TREND_SCORING_PARAMETER_SET}'"),
         nullable=False,
     )
-    baseline_log_odds: Mapped[float] = mapped_column(Numeric(10, 6), nullable=False)
-    starting_log_odds: Mapped[float] = mapped_column(Numeric(10, 6), nullable=False)
-    current_log_odds: Mapped[float] = mapped_column(Numeric(10, 6), nullable=False)
+    baseline_log_odds: Mapped[Decimal] = mapped_column(Numeric(10, 6), nullable=False)
+    starting_log_odds: Mapped[Decimal] = mapped_column(Numeric(10, 6), nullable=False)
+    current_log_odds: Mapped[Decimal] = mapped_column(Numeric(10, 6), nullable=False)
     decay_half_life_days: Mapped[int] = mapped_column(Integer, nullable=False)
     actor: Mapped[str | None] = mapped_column(String(255))
     context: Mapped[str | None] = mapped_column(String(255))
